+++ modules/hdf5/builtin/cpp/h5writeBuiltin.cpp
@@ -16,21 +16,42 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "h5infoBuiltin.hpp"
+#include "h5writeBuiltin.hpp"
 #include "Error.hpp"
+#include "HDF5_helpers.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
+// h5write(filename, datasetname, data)
+// h5write(filename, datasetname, data, start, count)
+// h5write(filename, datasetname, data, start, count, stride) 
+//=============================================================================
 ArrayOfVector
-Nelson::Hdf5Gateway::h5infoBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn)
+Nelson::Hdf5Gateway::h5writeBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn)
 {
     ArrayOfVector retval;
-    if (nLhs > 0) {
+    if (nLhs != 0) {
         Error(ERROR_WRONG_NUMBERS_OUTPUT_ARGS);
     }
-    if (argIn.size() < 2) {
+    indexType nbArgIn = argIn.size();
+    if (nbArgIn < 3 || nbArgIn > 6) {
         Error(ERROR_WRONG_NUMBERS_INPUT_ARGS);
     }
-    return retval;
+	boost::container::vector<double> start;
+    if (nbArgIn > 3) {
+    }
+    boost::container::vector<double> count;
+	if (nbArgIn > 4) {
+    }
+    boost::container::vector<double> stride;
+	if (nbArgIn > 5) {
+    }
+    ArrayOf param1 = argIn[0];
+    std::wstring filename = param1.getContentAsWideString();
+    ArrayOf param2 = argIn[1];
+    std::wstring dataSetName = param2.getContentAsWideString();
+    ArrayOf data = argIn[2];
+    hdf5Write(filename, dataSetName, data, start, count, stride);
+	return retval;
 }
 //=============================================================================