+++ modules/hdf5/src/c/dllMain.c
@@ -22,6 +22,9 @@
 #pragma comment(lib, CAT_3_STRINGS("boost_system-", BOOST_TARGET, ".lib"))
 #pragma comment(lib, CAT_3_STRINGS("boost_filesystem-", BOOST_TARGET, ".lib"))
 //=============================================================================
+#pragma comment(lib, "hdf5.lib")
+#pragma comment(lib, "hdf5_cpp.lib")
+//=============================================================================
 int WINAPI
 DllMain(HINSTANCE hInstance, DWORD reason, PVOID pvReserved)
 {